- move testUtil - add roundTrip (register, enroll, access "Single User")
