+++ src/test/scala/de/upb/cs/uc4/hyperledger/testBase/TestBase.scala
@@ -4,13 +4,13 @@ import java.nio.file.Path
 
 import de.upb.cs.uc4.hyperledger.connections.cases.{ ConnectionCertificate, ConnectionMatriculation }
 import de.upb.cs.uc4.hyperledger.connections.traits.{ ConnectionCertificateTrait, ConnectionMatriculationTrait }
-import de.upb.cs.uc4.hyperledger.utilities.{ EnrollmentManager, WalletManager }
 import de.upb.cs.uc4.hyperledger.utilities.helper.Logger
+import de.upb.cs.uc4.hyperledger.utilities.{ EnrollmentManager, WalletManager }
 
 class TestBase extends TestBaseTrait {
   private val testBase: TestBaseTrait = tryRetrieveEnvVar("Target") match {
     case "ProductionNetwork" => new TestBaseProductionNetwork
-    case _                   => new TestBaseDevNetwork
+    case _ => new TestBaseDevNetwork
   }
   override val networkDescriptionPath: Path = testBase.networkDescriptionPath
   override val caURL: String = testBase.caURL
@@ -21,54 +21,46 @@ class TestBase extends TestBaseTrait {
   override val channel: String = testBase.channel
   override val chaincode: String = testBase.chaincode
 
-  private def tryEnrollment(
-      caURL: String,
-      caCert: Path,
-      walletPath: Path,
-      enrollmentID: String,
-      enrollmentSecret: String,
-      organisationId: String,
-      channel: String,
-      chaincode: String,
-      networkDescriptionPath: Path
-  ): Unit = {
-    try {
-      debug("Try enrollment with: "
-        + " " + caURL
-        + " " + caCert
-        + " " + walletPath
-        + " " + enrollmentID
-        + " " + enrollmentSecret
-        + " " + organisationId)
+  protected[hyperledger] def tryEnrollment(
+                                            caURL: String,
+                                            caCert: Path,
+                                            walletPath: Path,
+                                            enrollmentID: String,
+                                            enrollmentSecret: String,
+                                            organisationId: String,
+                                            channel: String,
+                                            chaincode: String,
+                                            networkDescriptionPath: Path
+                                          ): Unit = {
+    debug("Try Enrollment")
+    if (!WalletManager.containsIdentity(this.walletPath, this.username)) {
+      try {
+        debug("Try enrollment with: "
+          + " " + caURL
+          + " " + caCert
+          + " " + walletPath
+          + " " + enrollmentID
+          + " " + enrollmentSecret
+          + " " + organisationId)
 
-      EnrollmentManager.enroll(caURL, caCert, walletPath, enrollmentID, enrollmentSecret, organisationId, channel, chaincode, networkDescriptionPath)
+        EnrollmentManager.enroll(caURL, caCert, walletPath, enrollmentID, enrollmentSecret, organisationId, channel, chaincode, networkDescriptionPath)
+      }
+      catch {
+        case e: Exception => Logger.warn("Enrollment failed, maybe some other test already enrolled the admin: " + e.getMessage)
+      }
     }
-    catch {
-      case e: Exception => Logger.warn("Enrollment failed, maybe some other test already enrolled the admin: " + e.getMessage)
-    }
+    debug("Finished Enrollment")
   }
 
   override def beforeAll(): Unit = {
     debug("Begin test with testBase Name = " + testBase.getClass.getName)
     if (testBase.isInstanceOf[TestBaseProductionNetwork]) {
-      while (!WalletManager.containsIdentity(this.walletPath, this.username)) {
-        tryEnrollment(
-          this.caURL,
-          this.tlsCert,
-          this.walletPath,
-          this.username,
-          this.password,
-          this.organisationId,
-          this.channel,
-          this.chaincode,
-          this.networkDescriptionPath
-        )
-      }
-      debug("Finished Enrollment")
+      tryEnrollment(this.caURL, this.tlsCert, this.walletPath, this.username, this.password, this.organisationId, this.channel, this.chaincode, this.networkDescriptionPath)
     }
   }
 
   def initializeMatriculation(userName: String = this.username): ConnectionMatriculationTrait = ConnectionMatriculation(userName, this.channel, this.chaincode, this.walletPath, this.networkDescriptionPath)
+
   def initializeCertificate(userName: String = this.username): ConnectionCertificateTrait = ConnectionCertificate(userName, this.channel, this.chaincode, this.walletPath, this.networkDescriptionPath)
 
   private def tryRetrieveEnvVar(varName: String, fallBack: String = ""): String = {
