+++ src/test/scala/de/upb/cs/uc4/hyperledger/tests/CertificateAccessTests.scala
@@ -2,7 +2,7 @@ package de.upb.cs.uc4.hyperledger.tests
 
 import de.upb.cs.uc4.hyperledger.connections.traits.ConnectionCertificateTrait
 import de.upb.cs.uc4.hyperledger.testBase.TestBase
-import de.upb.cs.uc4.hyperledger.testUtil.TestHelper
+import de.upb.cs.uc4.hyperledger.tests.testUtil.TestHelper
 
 class CertificateAccessTests extends TestBase {
